+++ marxs/simulator/tests/test_parallel.py
@@ -1,6 +1,7 @@
 # Licensed under GPL version 3 - see LICENSE.rst
 import numpy as np
 from transforms3d.affines import decompose
+from transforms3d.euler import euler2mat
 
 from ...utils import generate_test_photons
 from .. import Parallel, ParallelCalculated
@@ -20,6 +21,41 @@ def test_Parallel_numbering():
     photons = generate_test_photons(5)
     photons = detect(photons)
     assert np.all(photons['CCD_ID'] == 3.)
+
+def test_Parallel_moving():
+    '''Test moving center of structure after generation.
+    '''
+    pos = [[0, -10.1, -10.1],[0, .1, -10.1],[0, -10.1, .1],[0, .1, .1]]
+    detkwargs = {'elem_class': CCD,
+                 'elem_args': {'pixsize': 0.01, 'zoom': 5},
+                 'elem_pos': {'position': pos},
+                 'id_col': 'CCD_ID',
+                 }
+    detect = Parallel(**detkwargs)
+    detect1 = Parallel(**detkwargs)
+
+    move = np.eye(4)
+    move[2, 3] = 15
+
+    detect1.move_center(move)
+
+    # Check that the default positions are the same
+    for i in range(len(detect.elements)):
+        assert np.allclose(detect.elements[i].pos4d,
+                           detect1.elements[i].pos4d)
+
+    # but rotations are applied around a different point
+    rot = np.eye(4)
+    rot[:3, :3] = euler2mat(.3, .2, 1.2)
+    detect.uncertainty = rot
+    detect.generate_elements()
+
+    detect1.uncertainty = rot
+    detect1.generate_elements()
+
+    for i in range(len(detect.elements)):
+        assert ~np.allclose(detect.elements[i].pos4d,
+                            detect1.elements[i].pos4d)
 
 def test_Parallel_offset_numbering():
     '''The randomly generated positions might generate overlapping elements,